Move signatures publication abstraction in signer
To a dedicated module.
